Move interview model logic to controller
The logic for loading and tracking the Passages and Questions into the
system (and eventually determining the order / script) should not be in
the App Delegate or even in the View Controllers.

Instead, as inspired by [this
article](https://code.tutsplus.com/tutorials/the-right-way-to-share-state-between-swift-view-controllers--cms-28474)
we're going to have a model controller responsible for that data
management.  The model controller will be a dependency that gets
injected from view to view, and used to render the script.

Issue #13 Set up dynamic passage rendering
Issue #14 Set up dynamic question script
